first connection with the database on the graphql side, we can query email, name and batch of a participant
